iot-edge/modules/sensehat/requirements.txt
azure-iot-device~=2.7.0